[Removed test file]
   - Deleted test_trial.cpp, which was created just to test
     googletest.
