+++ src/test/CMakeLists.txt
@@ -5,6 +5,6 @@ add_subdirectory(lib/googletest)
 include_directories(lib/googletest/googletest/include)
 include_directories(lib/googletest/googlemock/include)
 
-add_executable(VDSProject_test main_test.cpp test_trial.cpp)
+add_executable(VDSProject_test main_test.cpp)
 target_link_libraries(VDSProject_test Manager)
 target_link_libraries(VDSProject_test gtest gtest_main pthread)
